requirements.txt
requests==2.18.4
pytz==2017.3
InstagramAPI==1.0.2
Pillow==5.0.0
PyYAML==3.12
